feat: Optimize and resolve Traccar bugs by migration to v5

+++ src/applications/reports/services/pdf/charts/use_of_vehicles/real_use_of_vehicle.py
@@ -88,5 +88,5 @@ class RealUseOfVehicleChart(ChartGenerator):
         )
 
     def get_stats(self):
-        return np.array(self.reserved_hours, np.float), \
-               np.array(self.free_hours, np.float)
+        return np.array(self.reserved_hours, float), \
+               np.array(self.free_hours, float)
